+++ HKMirror/Box/uLongBox.cs
@@ -98,5 +98,5 @@ public sealed class uLongBox : IComparable, IFormattable, IConvertible, ICompara
     public decimal ToDecimal(IFormatProvider provider) => Convert.ToDecimal(Value);
     public DateTime ToDateTime(IFormatProvider provider) => throw new InvalidCastException($"Invalid cast from '{nameof(uLongBox)}' to '{nameof(DateTime)}'.");
     public string ToString(IFormatProvider provider) => Value.ToString(provider);
-    public object ToType(Type type, IFormatProvider provider) => (object) Convert.ToInt32(Value, provider);
+    public object ToType(Type type, IFormatProvider provider) => (object) Convert.ToUInt64(Value, provider);
 }
